Split MCPServerProxy.swift into extension files
Goal: bring the file under file_length (1000) and the actor body under
type_body_length (350). Before: 1788-line file with a 1456-line actor
body. After: 157-line main file (just properties, nested types,
didSet observers, init), and 9 extension files grouped by concern:

- MCPServerProxy+API.swift — public list/read/get/call/ping API
- MCPServerProxy+Lifecycle.swift — connect/disconnect/initialize/
  startLineConnection/sendNotification/handleStreamTermination
- MCPServerProxy+IncomingMessages.swift — processIncomingMessage,
  handleNotification, handlePingRequest, log/progress notification
  formatting
- MCPServerProxy+NotificationHandlers.swift — public setters and
  setNotificationHandler<Payload>/removeNotificationHandler
- MCPServerProxy+NotificationRegistration.swift — internal update*
  routines + buildClientCapabilities + decodeNotificationPayload
- MCPServerProxy+ResponseHandling.swift — log/percent/text-extraction
  helpers + requestResult + errorMessage + decodeJSONPayload +
  responseMessage
- MCPServerProxy+Send.swift — send + sendLineMessage/sendSSEMessage/
  dispatchSSESend
- MCPServerProxy+SSE.swift — URL helpers + connectSSE Apple/Linux +
  shared SSE loops
- MCPServerProxy+StreamableHTTP.swift — sendStreamableRequest{Apple,
  Linux} + streamableRequestResponse* + readStreamableSSE*

Access widening from private → internal (since `private` doesn't cross
file boundaries): NotificationHandlerBox, every nested Result/Params
type, NotificationMethod, cachedTools, nextRequestID, responseTasks,
streamFailure, isDisconnecting, notificationHandlers, streamTask,
lineConnection, endpointContinuation. `endpointURL`, `sessionID`, and
the four `serverXxx` properties moved from `private(set)` to
`internal(set)` so extension setters work. Also `JSONRPCID.stringValue`
widened to internal.

Build clean. swift test --filter MCPServerProxyTests passes (13 tests).
