+++ Sources/SwiftMCP/Client/MCPServerProxy.swift
@@ -6,32 +6,39 @@ import Logging
 
 /// A proxy for interacting with an MCP server over stdio, TCP, or SSE.
 public final actor MCPServerProxy {
-    private struct NotificationHandlerBox: Sendable {
+    internal struct NotificationHandlerBox: Sendable {
         let payloadTypeDescription: String
-        let handle: @Sendable (MCPServerProxy, JSONRPCMessage.JSONRPCNotificationData) async throws -> Void
+        let handle: @Sendable (
+            MCPServerProxy,
+            JSONRPCMessage.JSONRPCNotificationData
+        ) async throws -> Void
     }
 
-    private struct ToolsListResult: Decodable, Sendable {
+    internal struct ToolsListResult: Decodable, Sendable {
         let tools: [MCPTool]
     }
 
-    private struct ResourcesListResult: Decodable, Sendable {
+    internal struct ResourcesListResult: Decodable, Sendable {
         let resources: [SimpleResource]
     }
 
-    private struct ResourceTemplatesListResult: Decodable, Sendable {
+    internal struct ResourceTemplatesListResult: Decodable, Sendable {
         let resourceTemplates: [SimpleResourceTemplate]
     }
 
-    private struct ResourceReadResult: Decodable, Sendable {
+    internal struct ResourceReadResult: Decodable, Sendable {
         let contents: [GenericResourceContent]
     }
 
-    private struct PromptsListResult: Decodable, Sendable {
+    internal struct PromptsListResult: Decodable, Sendable {
         let prompts: [Prompt]
     }
 
-    private enum NotificationMethod {
+    internal struct ResourceUpdatedParams: Decodable, Sendable {
+        let uri: String
+    }
+
+    internal enum NotificationMethod {
         static let log = "notifications/message"
         static let progress = "notifications/progress"
         static let toolsListChanged = "notifications/tools/list_changed"
@@ -54,28 +61,28 @@ public final actor MCPServerProxy {
     /// Base metadata included in _meta for ALL requests (e.g., accessToken).
     public var meta: JSONDictionary = [:]
 
-    private var cachedTools: [MCPTool]?
+    internal var cachedTools: [MCPTool]?
     private var requestIdSequence: Int = 0
 
-    private func nextRequestID() -> JSONRPCID {
+    internal func nextRequestID() -> JSONRPCID {
         defer { requestIdSequence += 1 }
         return .int(requestIdSequence)
     }
 
-    private var responseTasks: [JSONRPCID: CheckedContinuation<JSONRPCMessage, Error>] = [:]
-    private var streamFailure: Error?
-    private var isDisconnecting = false
+    internal var responseTasks: [JSONRPCID: CheckedContinuation<JSONRPCMessage, Error>] = [:]
+    internal var streamFailure: Error?
+    internal var isDisconnecting = false
 
-    public private(set) var endpointURL: URL?
-    public private(set) var sessionID: String?
-    private var streamTask: Task<Void, Error>?
+    public internal(set) var endpointURL: URL?
+    public internal(set) var sessionID: String?
+    internal var streamTask: Task<Void, Error>?
 
-    public private(set) var serverName: String?
-    public private(set) var serverVersion: String?
-    public private(set) var serverDescription: String?
-    public private(set) var serverCapabilities: ServerCapabilities?
+    public internal(set) var serverName: String?
+    public internal(set) var serverVersion: String?
+    public internal(set) var serverDescription: String?
+    public internal(set) var serverCapabilities: ServerCapabilities?
 
-    private var notificationHandlers: [String: NotificationHandlerBox] = [:]
+    internal var notificationHandlers: [String: NotificationHandlerBox] = [:]
 
     /// Optional handler for log notifications from the server.
     public var logNotificationHandler: (any MCPServerProxyLogNotificationHandling)? {
@@ -91,16 +98,6 @@ public final actor MCPServerProxy {
         }
     }
 
-    /// Updates the log notification handler.
-    public func setLogNotificationHandler(_ handler: (any MCPServerProxyLogNotificationHandling)?) {
-        logNotificationHandler = handler
-    }
-
-    /// Updates the progress notification handler.
-    public func setProgressNotificationHandler(_ handler: (any MCPServerProxyProgressNotificationHandling)?) {
-        progressNotificationHandler = handler
-    }
-
     /// Optional handler for resource-updated notifications from the server.
     /// Installing this handler enables the client to receive `notifications/resources/updated`
     /// for subscribed resources. The client automatically advertises subscription support
@@ -111,1672 +108,45 @@ public final actor MCPServerProxy {
         }
     }
 
-    /// Updates the resource notification handler.
-    public func setResourceNotificationHandler(_ handler: (any MCPServerProxyResourceNotificationHandling)?) {
-        resourceNotificationHandler = handler
-    }
-
     /// Optional handler for tools list-changed notifications.
-    /// When set, the client advertises support during initialization (if set before `connect()`).
+    /// When set, the client advertises support during initialization
+    /// (if set before `connect()`).
     public var toolsListChangedHandler: (any MCPServerProxyToolsListChangedHandling)? {
         didSet {
             updateToolsListChangedRegistration()
         }
     }
 
-    /// Updates the tools list-changed handler.
-    public func setToolsListChangedHandler(_ handler: (any MCPServerProxyToolsListChangedHandling)?) {
-        toolsListChangedHandler = handler
-    }
-
     /// Optional handler for resources list-changed notifications.
-    /// When set, the client advertises support during initialization (if set before `connect()`).
+    /// When set, the client advertises support during initialization
+    /// (if set before `connect()`).
     public var resourcesListChangedHandler: (any MCPServerProxyResourcesListChangedHandling)? {
         didSet {
             updateResourcesListChangedRegistration()
         }
     }
 
-    /// Updates the resources list-changed handler.
-    public func setResourcesListChangedHandler(_ handler: (any MCPServerProxyResourcesListChangedHandling)?) {
-        resourcesListChangedHandler = handler
-    }
-
     /// Optional handler for prompts list-changed notifications.
-    /// When set, the client advertises support during initialization (if set before `connect()`).
+    /// When set, the client advertises support during initialization
+    /// (if set before `connect()`).
     public var promptsListChangedHandler: (any MCPServerProxyPromptsListChangedHandling)? {
         didSet {
             updatePromptsListChangedRegistration()
         }
     }
 
-    /// Updates the prompts list-changed handler.
-    public func setPromptsListChangedHandler(_ handler: (any MCPServerProxyPromptsListChangedHandling)?) {
-        promptsListChangedHandler = handler
-    }
-
-    /// Registers a typed handler for a JSON-RPC notification.
-    public func setNotificationHandler<Payload>(
-        _ method: String,
-        as payloadType: Payload.Type = Payload.self,
-        handler: @escaping @Sendable (Payload) async -> Void
-    ) where Payload: Decodable, Payload: Sendable {
-        notificationHandlers[method] = NotificationHandlerBox(
-            payloadTypeDescription: String(reflecting: payloadType),
-            handle: { _, notification in
-                let payload = try Self.decodeNotificationPayload(from: notification, as: payloadType)
-                await handler(payload)
-            }
-        )
-    }
-
-    /// Removes the registered handler for a JSON-RPC notification.
-    public func removeNotificationHandler(for method: String) {
-        notificationHandlers.removeValue(forKey: method)
-    }
-
-    private var lineConnection: (any StdioConnection)?
-    private var endpointContinuation: CheckedContinuation<URL, Error>?
+    internal var lineConnection: (any StdioConnection)?
+    internal var endpointContinuation: CheckedContinuation<URL, Error>?
 
     public init(config: MCPServerConfig, cacheToolsList: Bool = false) {
         self.config = config
         self.service = nil
         self.cacheToolsList = cacheToolsList
     }
-
-    /// Connects to the MCP server and establishes an SSE, TCP, or stdio connection.
-    /// - Parameters:
-    ///   - clientName: The client name sent during the MCP handshake.
-    ///   - clientVersion: The client version sent during the MCP handshake.
-    public func connect(clientName: String = "swiftmcp-client", clientVersion: String = "1.0.0") async throws {
-        isDisconnecting = false
-        streamFailure = nil
-        endpointURL = nil
-
-        switch config {
-        case .stdio(let stdioConfig):
-            sessionID = UUID().uuidString
-            lineConnection = MCPServerProcess(config: stdioConfig)
-            try await startLineConnection()
-            try await initialize(clientName: clientName, clientVersion: clientVersion)
-
-        case .stdioHandles(let server):
-            sessionID = UUID().uuidString
-            lineConnection = InProcessStdioBridge(server: server)
-            try await startLineConnection()
-            try await initialize(clientName: clientName, clientVersion: clientVersion)
-
-        case .tcp(let tcpConfig):
-            #if canImport(Network)
-                sessionID = UUID().uuidString
-                let resolvedConfig = resolveTcpConfig(tcpConfig)
-                lineConnection = TCPConnection(config: resolvedConfig)
-                try await startLineConnection()
-                try await initialize(clientName: clientName, clientVersion: clientVersion)
-            #else
-                throw MCPServerProxyError.unsupportedPlatform("TCP connections require the Network framework.")
-            #endif
-
-        case .sse(let sseConfig):
-            #if os(Linux)
-                try await connectSSELinux(sseConfig: sseConfig, clientName: clientName, clientVersion: clientVersion)
-            #else
-                if #available(macOS 12.0, iOS 15.0, tvOS 15.0, watchOS 8.0, macCatalyst 15.0, *) {
-                    try await connectSSEApple(sseConfig: sseConfig, clientName: clientName, clientVersion: clientVersion)
-                } else {
-                    throw MCPServerProxyError.unsupportedPlatform("SSE client connections require macOS 12.0 or newer.")
-                }
-            #endif
-        }
-    }
-
-    /// Disconnects from the MCP server.
-    public func disconnect() async {
-        isDisconnecting = true
-        let disconnectError = CancellationError()
-        streamFailure = disconnectError
-        failPendingResponseTasks(with: disconnectError)
-        if let endpointContinuation {
-            self.endpointContinuation = nil
-            endpointContinuation.resume(throwing: disconnectError)
-        }
-
-        streamTask?.cancel()
-        streamTask = nil
-        endpointURL = nil
-
-        switch config {
-        case .stdio, .stdioHandles, .tcp:
-            await lineConnection?.stop()
-            lineConnection = nil
-        case .sse:
-            break
-        }
-
-        sessionID = nil
-    }
-
-    /// Lists all available tools from the server.
-    public func listTools() async throws -> [MCPTool] {
-        if cacheToolsList, let cachedTools = cachedTools {
-            return cachedTools
-        }
-
-        let result: ToolsListResult = try await requestResult(method: "tools/list", as: ToolsListResult.self)
-        let tools = result.tools
-
-        if cacheToolsList {
-            cachedTools = tools
-        }
-
-        return tools
-    }
-
-    /// Lists all static resources available from the server.
-    public func listResources() async throws -> [SimpleResource] {
-        let result: ResourcesListResult = try await requestResult(method: "resources/list", as: ResourcesListResult.self)
-        return result.resources
-    }
-
-    /// Lists all resource templates available from the server.
-    public func listResourceTemplates() async throws -> [SimpleResourceTemplate] {
-        let result: ResourceTemplatesListResult = try await requestResult(method: "resources/templates/list", as: ResourceTemplatesListResult.self)
-        return result.resourceTemplates
-    }
-
-    /// Reads a resource at the specified URI.
-    public func readResource(uri: URL) async throws -> [GenericResourceContent] {
-        let result: ResourceReadResult = try await requestResult(
-            method: "resources/read",
-            params: ["uri": .string(uri.absoluteString)],
-            as: ResourceReadResult.self
-        )
-        return result.contents
-    }
-
-    /// Subscribes to update notifications for a resource URI.
-    /// Requires a `resourceNotificationHandler` to be set and the server
-    /// to advertise `resources.subscribe` capability.
-    public func subscribeResource(uri: URL) async throws {
-        _ = try await requestResult(
-            method: "resources/subscribe",
-            params: ["uri": .string(uri.absoluteString)]
-        )
-    }
-
-    /// Unsubscribes from update notifications for a resource URI.
-    public func unsubscribeResource(uri: URL) async throws {
-        _ = try await requestResult(
-            method: "resources/unsubscribe",
-            params: ["uri": .string(uri.absoluteString)]
-        )
-    }
-
-    /// Lists all prompts available from the server.
-    public func listPrompts() async throws -> [Prompt] {
-        let result: PromptsListResult = try await requestResult(method: "prompts/list", as: PromptsListResult.self)
-        return result.prompts
-    }
-
-    /// Gets a prompt by name with optional arguments.
-    public func getPrompt(
-        name: String,
-        arguments: JSONDictionary = [:]
-    ) async throws -> PromptResult {
-        try await requestResult(
-            method: "prompts/get",
-            params: [
-                "name": .string(name),
-                "arguments": .object(arguments)
-            ],
-            as: PromptResult.self
-        )
-    }
-
-    /// Calls a tool by name on the connected MCP server with the provided arguments.
-    public func callTool(
-        _ name: String,
-        arguments: JSONDictionary = [:],
-        progressToken: JSONValue? = .string(UUID().uuidString)
-    ) async throws -> String {
-        let requestId = nextRequestID()
-        var params: JSONDictionary = [
-            "name": .string(name),
-            "arguments": .object(arguments)
-        ]
-
-        // Merge base meta with progressToken
-        var requestMeta = meta  // Start with base meta (e.g., accessToken)
-        if let progressToken {
-            requestMeta["progressToken"] = progressToken
-        }
-
-        if !requestMeta.isEmpty {
-            params["_meta"] = .object(requestMeta)
-        }
-
-        let request = JSONRPCMessage.request(id: requestId, method: "tools/call", params: params)
-        let responseMessage = try await send(request)
-
-        let result: JSONDictionary
-        switch responseMessage {
-        case .response(let responseData):
-            guard let responseResult = responseData.result else {
-                throw MCPServerProxyError.communicationError("Invalid response type for tools/call, expected JSONRPCResponse")
-            }
-            result = responseResult
-        case .errorResponse(let errorResponse):
-            throw MCPServerProxyError.toolError(errorResponse.error.message)
-        default:
-            throw MCPServerProxyError.communicationError("Invalid response type for tools/call, expected JSONRPCResponse")
-        }
-
-        if result["isError"]?.boolValue == true {
-            throw MCPServerProxyError.toolError(errorMessage(from: result) ?? "Tool call failed with an unspecified error.")
-        }
-
-        guard let contentArray = result["content"]?.arrayValue else {
-            throw MCPServerProxyError.communicationError("Invalid content format in tools/call response")
-        }
-
-        if let text = extractTextPayload(from: contentArray) {
-            return text
-        }
-
-        if let contentPayload = encodeContentPayload(from: contentArray) {
-            return contentPayload
-        }
-
-        throw MCPServerProxyError.communicationError("Failed to extract string content from tools/call response")
-    }
-
-    /// Invalidates the cached list of tools.
-    public func invalidateToolsCache() {
-        cachedTools = nil
-    }
-
-    /// Sends a JSON-RPC message to the server and waits for the response.
-    public func send(_ message: JSONRPCMessage) async throws -> JSONRPCMessage {
-        let messageId = message.id
-        switch config {
-        case .stdio, .stdioHandles, .tcp:
-            guard let messageId = messageId else {
-                throw MCPServerProxyError.communicationError("Message must have an ID")
-            }
-            let encoder = JSONEncoder()
-            encoder.outputFormatting = [.sortedKeys]
-            let data = try encoder.encode(message)
-
-            let messageWithNewline = data + Data("\n".utf8)
-            guard let lineConnection else {
-                throw MCPServerProxyError.communicationError("Not connected to line-based server")
-            }
-
-            if let streamFailure {
-                throw streamFailure
-            }
-
-            await lineConnection.write(messageWithNewline)
-
-            if let streamFailure {
-                throw streamFailure
-            }
-
-            return try await withCheckedThrowingContinuation { (continuation: CheckedContinuation<JSONRPCMessage, Error>) in
-                if let streamFailure {
-                    continuation.resume(throwing: streamFailure)
-                } else {
-                    responseTasks[messageId] = continuation
-                }
-            }
-
-        case .sse(let sseConfig):
-            if isStreamableMCPURL(endpointURL ?? sseConfig.url) {
-                #if os(Linux)
-                    return try await sendStreamableRequestLinux(message, sseConfig: sseConfig)
-                #else
-                    if #available(macOS 12.0, iOS 15.0, tvOS 15.0, watchOS 8.0, macCatalyst 15.0, *) {
-                        return try await sendStreamableRequestApple(message, sseConfig: sseConfig)
-                    } else {
-                        throw MCPServerProxyError.unsupportedPlatform("Streamable HTTP requires macOS 12.0 or newer.")
-                    }
-                #endif
-            }
-
-            guard let endpointURL = endpointURL else {
-                throw MCPServerProxyError.communicationError("Not connected to server")
-            }
-            if let streamFailure {
-                throw streamFailure
-            }
-            guard let messageId = messageId else {
-                throw MCPServerProxyError.communicationError("Message must have an ID")
-            }
-            let session = URLSession(configuration: .default)
-            var urlRequest = URLRequest(url: endpointURL)
-            urlRequest.httpMethod = "POST"
-            urlRequest.setValue("application/json", forHTTPHeaderField: "Content-Type")
-            configureSSEPOSTRequest(&urlRequest, sseConfig: sseConfig)
-            let encoder = JSONEncoder()
-            encoder.outputFormatting = [.sortedKeys]
-            let data = try encoder.encode(message)
-            urlRequest.httpBody = data
-
-            return try await withCheckedThrowingContinuation { (continuation: CheckedContinuation<JSONRPCMessage, Error>) in
-                if let streamFailure {
-                    continuation.resume(throwing: streamFailure)
-                    return
-                }
-
-                responseTasks[messageId] = continuation
-
-                Task {
-                    do {
-                        let (responseData, response) = try await session.data(for: urlRequest)
-                        guard let httpResponse = response as? HTTPURLResponse else {
-                            throw MCPServerProxyError.communicationError("Invalid HTTP response")
-                        }
-
-                        switch httpResponse.statusCode {
-                        case 200, 202:
-                            if let updatedSessionID = httpResponse.value(forHTTPHeaderField: "Mcp-Session-Id") {
-                                sessionID = updatedSessionID
-                            }
-
-                            // Some servers reply immediately in the HTTP body even for 202.
-                            if let responseMessage = try responseMessage(for: messageId, from: responseData) {
-                                if responseTasks[messageId] != nil {
-                                    responseTasks.removeValue(forKey: messageId)
-                                    continuation.resume(returning: responseMessage)
-                                }
-                                return
-                            }
-
-                            if httpResponse.statusCode == 202 {
-                                return // response will arrive over SSE
-                            }
-
-                            guard let responseMessage = try responseMessage(for: messageId, from: responseData) else {
-                                let responseBody = String(data: responseData, encoding: .utf8)?.trimmingCharacters(in: .whitespacesAndNewlines) ?? ""
-                                let details = responseBody.isEmpty ? "" : ": \(responseBody)"
-                                throw MCPServerProxyError.communicationError("HTTP 200 did not include JSON-RPC response for request \(messageId.stringValue)\(details)")
-                            }
-
-                            if responseTasks[messageId] != nil {
-                                responseTasks.removeValue(forKey: messageId)
-                                continuation.resume(returning: responseMessage)
-                            }
-                        default:
-                            let responseBody = String(data: responseData, encoding: .utf8)?.trimmingCharacters(in: .whitespacesAndNewlines) ?? ""
-                            let details = responseBody.isEmpty ? "" : ": \(responseBody)"
-                            throw MCPServerProxyError.communicationError("HTTP error \(httpResponse.statusCode)\(details)")
-                        }
-                    } catch {
-                        if responseTasks[messageId] != nil {
-                            responseTasks.removeValue(forKey: messageId)
-                            continuation.resume(throwing: error)
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    /// Sends a ping request to the server.
-    public func ping() async throws {
-        let requestId = nextRequestID()
-        let request = JSONRPCMessage.request(id: requestId, method: "ping", params: nil)
-        _ = try await send(request)
-    }
-
-    private func startLineConnection() async throws {
-        guard let lineConnection else {
-            throw MCPServerProxyError.communicationError("Not connected to line-based server")
-        }
-        try await lineConnection.start()
-
-        streamTask = Task {
-            do {
-                let lines = await lineConnection.lines()
-                for try await data in lines {
-                    await processIncomingMessage(data: data)
-                }
-                self.handleStreamTermination(
-                    MCPServerProxyError.communicationError(
-                        "Connection closed by server before response was received"
-                    )
-                )
-            } catch is CancellationError {
-                // Pending requests are cancelled in disconnect().
-            } catch {
-                logger.error("[MCP DEBUG] Stream error: \(error.localizedDescription)")
-                self.handleStreamTermination(error)
-            }
-        }
-    }
-
-    // MARK: - Internal
-
-    private func failPendingResponseTasks(with error: Error) {
-        let pending = responseTasks
-        responseTasks.removeAll()
-
-        for (_, continuation) in pending {
-            continuation.resume(throwing: error)
-        }
-    }
-
-    private func handleStreamTermination(_ error: Error) {
-        guard !isDisconnecting else {
-            return
-        }
-
-        let failure: Error
-        if let serverError = error as? MCPServerProxyError {
-            failure = serverError
-        } else {
-            failure = MCPServerProxyError.communicationError(error.localizedDescription)
-        }
-
-        if streamFailure == nil {
-            streamFailure = failure
-        }
-
-        let terminalError = streamFailure ?? failure
-        failPendingResponseTasks(with: terminalError)
-
-        if let endpointContinuation {
-            self.endpointContinuation = nil
-            endpointContinuation.resume(throwing: terminalError)
-        }
-    }
-
-    private func initialize(clientName: String, clientVersion: String) async throws {
-        let requestId = nextRequestID()
-        var params: JSONDictionary = [
-            "protocolVersion": .string(HTTPSSETransport.latestProtocolVersion),
-            "clientInfo": .object([
-                "name": .string(clientName),
-                "version": .string(clientVersion)
-            ]),
-            "capabilities": .object(buildClientCapabilities())
-        ]
-
-        // Add base metadata if present
-        if !meta.isEmpty {
-            params["_meta"] = .object(meta)
-        }
-
-        let request = JSONRPCMessage.request(id: requestId, method: "initialize", params: params)
-        let response = try await send(request)
-
-        guard case let .response(responseData) = response,
-              let result = responseData.result else {
-            throw MCPServerProxyError.communicationError("Invalid initialize response")
-        }
-
-        let rawServerDescription = extractServerDescription(from: result)
-        let initResult = try Self.decodeJSONPayload(result, as: InitializeResult.self)
-        serverName = initResult.serverInfo.name
-        serverVersion = initResult.serverInfo.version
-        serverDescription = initResult.serverInfo.description ?? rawServerDescription
-        serverCapabilities = initResult.capabilities
-        if service == nil {
-            service = serverName
-        }
-        logger.info("Connected to MCP server: \(serverName ?? "unknown") version \(serverVersion ?? "unknown")")
-
-        // Send the required notifications/initialized to complete the MCP handshake.
-        // Servers may ignore subsequent requests until this notification is received.
-        try await sendNotification(JSONRPCMessage.notification(method: "notifications/initialized"))
-    }
-
-    /// Sends a JSON-RPC notification (fire-and-forget, no response expected).
-    private func sendNotification(_ message: JSONRPCMessage) async throws {
-        let encoder = JSONEncoder()
-        encoder.outputFormatting = [.sortedKeys]
-        let data = try encoder.encode(message)
-
-        switch config {
-        case .stdio, .stdioHandles, .tcp:
-            guard let lineConnection else {
-                throw MCPServerProxyError.communicationError("Not connected to line-based server")
-            }
-            await lineConnection.write(data + Data("\n".utf8))
-
-        case .sse(let sseConfig):
-            guard let url = endpointURL ?? (isStreamableMCPURL(sseConfig.url) ? sseConfig.url : nil) else {
-                throw MCPServerProxyError.communicationError("Not connected to server")
-            }
-            var request = URLRequest(url: url)
-            request.httpMethod = "POST"
-            request.httpBody = data
-            request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-            configureSSEPOSTRequest(&request, sseConfig: sseConfig)
-            let (_, response) = try await URLSession.shared.data(for: request)
-            if let httpResponse = response as? HTTPURLResponse,
-               !(200...299).contains(httpResponse.statusCode) {
-                throw MCPServerProxyError.communicationError(
-                    "Notification rejected by server (HTTP \(httpResponse.statusCode))"
-                )
-            }
-        }
-    }
-
-    private func resolveTcpConfig(_ config: MCPServerTcpConfig) -> MCPServerTcpConfig {
-        guard case .bonjour(let serviceName, let domain) = config.endpoint,
-              serviceName == nil,
-              let service else {
-            return config
-        }
-        return MCPServerTcpConfig(
-            serviceName: service,
-            domain: domain,
-            serviceType: config.serviceType,
-            timeout: config.timeout,
-            preferIPv4: config.preferIPv4
-        )
-    }
-
-    private func processIncomingMessage(event: String = "", data: String) async {
-        if data.isEmpty {
-            return
-        }
-
-        if event == "endpoint" {
-            if case .sse(let sseConfig) = config, isStreamableMCPURL(sseConfig.url) {
-                // Ignore legacy endpoint events for streamable /mcp mode.
-                return
-            }
-            if let url = URL(string: data) {
-                endpointURL = url
-                if let continuation = endpointContinuation {
-                    endpointContinuation = nil
-                    continuation.resume(returning: url)
-                }
-                return
-            }
-        }
-
-        guard let jsonData = data.data(using: .utf8) else { return }
-        logger.trace("[MCP DEBUG] Received JSON-RPC message: \(data)")
-        let decoder = JSONDecoder()
-        if let message = try? decoder.decode(JSONRPCMessage.self, from: jsonData) {
-            switch message {
-            case .request(let requestData):
-                if requestData.method == "ping" {
-                    logger.info("[MCP] Ping request received; sending response.")
-                    await handlePingRequest(requestData)
-                } else {
-                    logger.debug("[MCP DEBUG] Ignoring client request: \(requestData.method)")
-                }
-            case .notification(let notification):
-                await handleNotification(notification)
-            case .response, .errorResponse:
-                if let id = message.id, let waitingContinuation = responseTasks[id] {
-                    responseTasks.removeValue(forKey: id)
-                    waitingContinuation.resume(returning: message)
-                } else {
-                    logger.error("[MCP DEBUG] No waiting continuation found for ID \(message.id?.stringValue ?? "nil")")
-                }
-            }
-        } else {
-            logger.error("[MCP DEBUG] Failed to decode JSON-RPC message")
-        }
-    }
-
-    func handleNotification(_ notification: JSONRPCMessage.JSONRPCNotificationData) async {
-        if let handler = notificationHandlers[notification.method] {
-            do {
-                try await handler.handle(self, notification)
-                return
-            } catch {
-                logger.error("[MCP] Failed to handle \(notification.method) as \(handler.payloadTypeDescription): \(String(describing: error))")
-            }
-        }
-
-        handleUnhandledNotification(notification)
-    }
-
-    private func handlePingRequest(_ request: JSONRPCMessage.JSONRPCRequestData) async {
-        guard let endpointURL = endpointURL else { return }
-        guard case .sse(let sseConfig) = config else { return }
-        let response = JSONRPCMessage.response(id: request.id, result: [:])
-        let sessionConfig = URLSessionConfiguration.default
-        let session = URLSession(configuration: sessionConfig)
-        var urlRequest = URLRequest(url: endpointURL)
-        urlRequest.httpMethod = "POST"
-        urlRequest.setValue("application/json", forHTTPHeaderField: "Content-Type")
-        configureSSEPOSTRequest(&urlRequest, sseConfig: sseConfig)
-        let encoder = JSONEncoder()
-        let data = try? encoder.encode(response)
-        urlRequest.httpBody = data
-        do {
-            let (_, response) = try await session.data(for: urlRequest)
-            if let httpResponse = response as? HTTPURLResponse, ![200, 202].contains(httpResponse.statusCode) {
-                return
-            }
-        } catch {
-            return
-        }
-    }
-
-    private func handleUnhandledNotification(_ notification: JSONRPCMessage.JSONRPCNotificationData) {
-        switch notification.method {
-        case NotificationMethod.progress:
-            logProgressNotification(notification)
-        case NotificationMethod.log:
-            logIncomingLogMessage(notification)
-        default:
-            logger.trace("[MCP DEBUG] Received notification: \(notification.method)")
-        }
-    }
-
-    private func logProgressNotification(_ notification: JSONRPCMessage.JSONRPCNotificationData) {
-        guard let params = notification.params else {
-            logger.info("[MCP] Progress notification received.")
-            return
-        }
-        let tokenValue = params["progressToken"]
-        let progressValue = numericValue(params["progress"])
-        let totalValue = numericValue(params["total"])
-        let messageValue = params["message"]?.stringValue
-
-        var parts: [String] = []
-        if let messageValue, !messageValue.isEmpty {
-            parts.append(messageValue)
-        }
-        if let progressValue {
-            if let percentText = progressPercentText(progressValue: progressValue, totalValue: totalValue) {
-                parts.append("progress \(percentText)")
-            } else if let totalValue {
-                parts.append("progress \(progressValue)/\(totalValue)")
-            } else {
-                parts.append("progress \(progressValue)")
-            }
-        }
-        if let tokenValue {
-            parts.append("token \(tokenValue)")
-        }
-
-        if parts.isEmpty {
-            logger.info("[MCP] Progress notification received.")
-        } else {
-            logger.info("[MCP] Progress: \(parts.joined(separator: " | "))")
-        }
-    }
-
-    func handleLogNotification(_ notification: JSONRPCMessage.JSONRPCNotificationData) async {
-        await handleNotification(notification)
-    }
-
-    private func updateLogNotificationRegistration() {
-        guard let handler = logNotificationHandler else {
-            removeNotificationHandler(for: NotificationMethod.log)
-            return
-        }
-
-        notificationHandlers[NotificationMethod.log] = NotificationHandlerBox(
-            payloadTypeDescription: String(reflecting: LogMessage.self),
-            handle: { proxy, notification in
-                let message = try Self.decodeNotificationPayload(from: notification, as: LogMessage.self)
-                await handler.mcpServerProxy(proxy, didReceiveLog: message)
-            }
-        )
-    }
-
-    private func updateProgressNotificationRegistration() {
-        guard let handler = progressNotificationHandler else {
-            removeNotificationHandler(for: NotificationMethod.progress)
-            return
-        }
-
-        notificationHandlers[NotificationMethod.progress] = NotificationHandlerBox(
-            payloadTypeDescription: String(reflecting: ProgressNotification.self),
-            handle: { proxy, notification in
-                let progress = try Self.decodeNotificationPayload(from: notification, as: ProgressNotification.self)
-                await handler.mcpServerProxy(proxy, didReceiveProgress: progress)
-            }
-        )
-    }
-
-    // MARK: - Resource Updated Registration
-
-    private struct ResourceUpdatedParams: Decodable, Sendable {
-        let uri: String
-    }
-
-    private func updateResourceNotificationRegistration() {
-        guard let handler = resourceNotificationHandler else {
-            removeNotificationHandler(for: NotificationMethod.resourceUpdated)
-            return
-        }
-
-        notificationHandlers[NotificationMethod.resourceUpdated] = NotificationHandlerBox(
-            payloadTypeDescription: "ResourceUpdated",
-            handle: { proxy, notification in
-                let params = try Self.decodeNotificationPayload(from: notification, as: ResourceUpdatedParams.self)
-                guard let url = URL(string: params.uri) else { return }
-                await handler.mcpServerProxy(proxy, resourceUpdatedAt: url)
-            }
-        )
-    }
-
-    // MARK: - List Changed Registration
-
-    private func updateToolsListChangedRegistration() {
-        guard let handler = toolsListChangedHandler else {
-            removeNotificationHandler(for: NotificationMethod.toolsListChanged)
-            return
-        }
-
-        notificationHandlers[NotificationMethod.toolsListChanged] = NotificationHandlerBox(
-            payloadTypeDescription: "ToolsListChanged",
-            handle: { proxy, _ in
-                await handler.mcpServerProxyToolsListDidChange(proxy)
-            }
-        )
-    }
-
-    private func updateResourcesListChangedRegistration() {
-        guard let handler = resourcesListChangedHandler else {
-            removeNotificationHandler(for: NotificationMethod.resourcesListChanged)
-            return
-        }
-
-        notificationHandlers[NotificationMethod.resourcesListChanged] = NotificationHandlerBox(
-            payloadTypeDescription: "ResourcesListChanged",
-            handle: { proxy, _ in
-                await handler.mcpServerProxyResourcesListDidChange(proxy)
-            }
-        )
-    }
-
-    private func updatePromptsListChangedRegistration() {
-        guard let handler = promptsListChangedHandler else {
-            removeNotificationHandler(for: NotificationMethod.promptsListChanged)
-            return
-        }
-
-        notificationHandlers[NotificationMethod.promptsListChanged] = NotificationHandlerBox(
-            payloadTypeDescription: "PromptsListChanged",
-            handle: { proxy, _ in
-                await handler.mcpServerProxyPromptsListDidChange(proxy)
-            }
-        )
-    }
-
-    // MARK: - Client Capabilities
-
-    /// Builds the client capabilities dictionary based on installed handlers.
-    /// Only advertises support for features that have handlers installed.
-    private func buildClientCapabilities() -> JSONDictionary {
-        var capabilities: JSONDictionary = [:]
-
-        // Resource capabilities — subscription and/or list-changed
-        var resourcesCap: JSONDictionary = [:]
-        if resourceNotificationHandler != nil {
-            resourcesCap["subscribe"] = .bool(true)
-        }
-        if resourcesListChangedHandler != nil {
-            resourcesCap["listChanged"] = .bool(true)
-        }
-        if !resourcesCap.isEmpty {
-            capabilities["resources"] = .object(resourcesCap)
-        }
-
-        // Tools list-changed
-        if toolsListChangedHandler != nil {
-            capabilities["tools"] = .object(["listChanged": .bool(true)])
-        }
-
-        // Prompts list-changed
-        if promptsListChangedHandler != nil {
-            capabilities["prompts"] = .object(["listChanged": .bool(true)])
-        }
-
-        return capabilities
-    }
-
-    private static func decodeNotificationPayload<Payload>(
-        from notification: JSONRPCMessage.JSONRPCNotificationData,
-        as payloadType: Payload.Type = Payload.self
-    ) throws -> Payload where Payload: Decodable {
-        let params = notification.params ?? [:]
-        return try decodeJSONPayload(params, as: payloadType)
-    }
-
-    private func logIncomingLogMessage(_ message: LogMessage) {
-        var parts: [String] = []
-        parts.append("level \(message.level.rawValue)")
-        if let loggerName = message.logger, !loggerName.isEmpty {
-            parts.append("logger \(loggerName)")
-        }
-        let dataDescription = String(describing: message.data)
-        if !dataDescription.isEmpty {
-            parts.append("data \(dataDescription)")
-        }
-
-        let level = loggerLevel(for: message.level)
-        if parts.isEmpty {
-            logger.log(level: level, "[MCP] Log notification received.")
-        } else {
-            logger.log(level: level, "[MCP] Log: \(parts.joined(separator: " | "))")
-        }
-    }
-
-    private func logIncomingLogMessage(_ notification: JSONRPCMessage.JSONRPCNotificationData) {
-        guard let params = notification.params else {
-            logger.info("[MCP] Log notification received.")
-            return
-        }
-
-        let levelValue = params["level"]?.stringValue
-        let level = levelValue.flatMap(LogLevel.init(string:)) ?? .info
-        let loggerName = params["logger"]?.stringValue
-        let dataValue = params["data"] ?? .string("")
-        let message = LogMessage(level: level, logger: loggerName, data: dataValue)
-        logIncomingLogMessage(message)
-    }
-
-    private func loggerLevel(for level: LogLevel) -> Logger.Level {
-        switch level {
-        case .debug:
-            return .debug
-        case .info:
-            return .info
-        case .notice:
-            return .notice
-        case .warning:
-            return .warning
-        case .error:
-            return .error
-        case .critical, .alert, .emergency:
-            return .critical
-        }
-    }
-
-    private func numericValue(_ value: JSONValue?) -> Double? {
-        value?.doubleValue
-    }
-
-    private func extractTextPayload(from contentArray: JSONArray) -> String? {
-        guard contentArray.count == 1 else {
-            return nil
-        }
-        for content in contentArray {
-            guard let contentDict = contentDictionary(from: content),
-                  let type = stringValue(contentDict["type"]),
-                  type == "text",
-                  let text = stringValue(contentDict["text"]) else {
-                continue
-            }
-            return text
-        }
-        return nil
-    }
-
-    private func encodeContentPayload(from contentArray: JSONArray) -> String? {
-        let normalized = contentArray.compactMap { contentDictionary(from: $0).map(JSONValue.object) }
-        if normalized.isEmpty {
-            return "[]"
-        }
-        let payload: JSONValue = normalized.count == 1 ? normalized[0] : .array(normalized)
-        let encoder = MCPJSONCoding.makeWireEncoder()
-        guard let data = try? encoder.encode(payload),
-              let string = String(data: data, encoding: .utf8) else {
-            return nil
-        }
-        return string
-    }
-
-    private func contentDictionary(from value: JSONValue) -> JSONDictionary? {
-        value.dictionaryValue
-    }
-
-    private func stringValue(_ value: JSONValue?) -> String? {
-        value?.stringValue
-    }
-
-    private func requestResult(
-        method: String,
-        params: JSONDictionary? = nil
-    ) async throws -> JSONDictionary {
-        let requestId = nextRequestID()
-        var requestParams = params ?? [:]
-
-        if !meta.isEmpty {
-            var requestMeta = meta
-            if let existingMeta = requestParams["_meta"]?.dictionaryValue {
-                requestMeta.merge(existingMeta) { _, new in new }
-            }
-            requestParams["_meta"] = .object(requestMeta)
-        }
-
-        let request = JSONRPCMessage.request(id: requestId, method: method, params: requestParams.isEmpty ? nil : requestParams)
-        let response = try await send(request)
-
-        switch response {
-        case .response(let responseData):
-            guard let result = responseData.result else {
-                throw MCPServerProxyError.communicationError("Invalid response type for \(method)")
-            }
-            if result["isError"]?.boolValue == true {
-                throw MCPServerProxyError.communicationError(errorMessage(from: result) ?? "Request failed for \(method)")
-            }
-            return result
-        case .errorResponse(let errorResponse):
-            throw MCPServerProxyError.communicationError(errorResponse.error.message)
-        default:
-            throw MCPServerProxyError.communicationError("Invalid response type for \(method)")
-        }
-    }
-
-    private func requestResult<T: Decodable>(
-        method: String,
-        params: JSONDictionary? = nil,
-        as type: T.Type = T.self
-    ) async throws -> T {
-        let result = try await requestResult(method: method, params: params)
-        return try result.decoded(type)
-    }
-
-    private static func decodeJSONPayload<T: Decodable, Payload: Encodable>(
-        _ payload: Payload,
-        as type: T.Type = T.self
-    ) throws -> T {
-        let encoder = MCPJSONCoding.makeWireEncoder()
-        let data = try encoder.encode(payload)
-
-        let decoder = MCPJSONCoding.makeDecoder()
-        return try decoder.decode(type, from: data)
-    }
-
-    private func errorMessage(from result: JSONDictionary) -> String? {
-        if let message = stringValue(result["message"]) {
-            return message
-        }
-        if let contentArray = result["content"]?.arrayValue {
-            return extractTextPayload(from: contentArray)
-        }
-        return nil
-    }
-
-    private func extractServerDescription(from result: JSONDictionary) -> String? {
-        result["serverInfo"]?.dictionaryValue?["description"]?.stringValue
-    }
-
-    private func progressPercentText(progressValue: Double, totalValue: Double?) -> String? {
-        if let totalValue, totalValue > 0 {
-            return formatPercent((progressValue / totalValue) * 100)
-        }
-        if progressValue >= 0, progressValue <= 1 {
-            return formatPercent(progressValue * 100)
-        }
-        return nil
-    }
-
-    private func formatPercent(_ percent: Double) -> String {
-        let rounded = Int(percent.rounded())
-        return "\(rounded)%"
-    }
-
-    private func isStreamableMCPURL(_ url: URL) -> Bool {
-        guard let components = URLComponents(url: url, resolvingAgainstBaseURL: false) else {
-            return false
-        }
-        return components.path.hasPrefix("/mcp")
-    }
-
-    private func applyConfiguredSSEHeaders(_ request: inout URLRequest, sseConfig: MCPServerSseConfig) {
-        for (key, value) in sseConfig.headers {
-            request.setValue(value, forHTTPHeaderField: key)
-        }
-
-        // Allow token in metadata to override auth header from config.
-        if let accessToken = meta["accessToken"]?.stringValue {
-            request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-        }
-    }
-
-    private func applyStreamableProtocolHeaders(_ request: inout URLRequest) {
-        request.setValue(HTTPSSETransport.latestProtocolVersion, forHTTPHeaderField: "MCP-Protocol-Version")
-        if let sessionID {
-            request.setValue(sessionID, forHTTPHeaderField: "Mcp-Session-Id")
-        }
-    }
-
-    private func configureSSEPOSTRequest(_ request: inout URLRequest, sseConfig: MCPServerSseConfig) {
-        applyConfiguredSSEHeaders(&request, sseConfig: sseConfig)
-        request.setValue("application/json, text/event-stream", forHTTPHeaderField: "Accept")
-
-        if isStreamableMCPURL(sseConfig.url) {
-            applyStreamableProtocolHeaders(&request)
-        } else if let sessionID {
-            request.setValue(sessionID, forHTTPHeaderField: "Mcp-Session-Id")
-        }
-    }
-
-    private func configureSSEGETRequest(
-        _ request: inout URLRequest,
-        sseConfig: MCPServerSseConfig,
-        lastEventID: String? = nil
-    ) {
-        applyConfiguredSSEHeaders(&request, sseConfig: sseConfig)
-        request.setValue("text/event-stream", forHTTPHeaderField: "Accept")
-
-        if isStreamableMCPURL(sseConfig.url) {
-            applyStreamableProtocolHeaders(&request)
-            if let lastEventID {
-                request.setValue(lastEventID, forHTTPHeaderField: "Last-Event-ID")
-            }
-        }
-    }
-
-    private func responseMessage(for requestID: JSONRPCID, from data: Data) throws -> JSONRPCMessage? {
-        guard !data.isEmpty else {
-            return nil
-        }
-
-        let messages = try JSONRPCMessage.decodeMessages(from: data)
-        return messages.first { message in
-            guard message.id == requestID else {
-                return false
-            }
-            switch message {
-            case .response, .errorResponse:
-                return true
-            case .request, .notification:
-                return false
-            }
-        }
-    }
-
-    #if !os(Linux)
-        @available(macOS 12.0, iOS 15.0, tvOS 15.0, watchOS 8.0, macCatalyst 15.0, *)
-        private func sendStreamableRequestApple(_ message: JSONRPCMessage, sseConfig: MCPServerSseConfig) async throws -> JSONRPCMessage {
-            guard let requestID = message.id else {
-                throw MCPServerProxyError.communicationError("Message must have an ID")
-            }
-            guard let endpointURL = endpointURL ?? (isStreamableMCPURL(sseConfig.url) ? sseConfig.url : nil) else {
-                throw MCPServerProxyError.communicationError("Not connected to server")
-            }
-
-            let encoder = JSONEncoder()
-            encoder.outputFormatting = [.sortedKeys]
-            let requestBody = try encoder.encode(message)
-
-            return try await streamableRequestResponseApple(
-                endpointURL: endpointURL,
-                sseConfig: sseConfig,
-                requestID: requestID,
-                requestBody: requestBody,
-                lastEventID: nil,
-                retryMilliseconds: 1000
-            )
-        }
-
-        @available(macOS 12.0, iOS 15.0, tvOS 15.0, watchOS 8.0, macCatalyst 15.0, *)
-        // swiftlint:disable:next function_parameter_count
-        private func streamableRequestResponseApple(
-            endpointURL: URL,
-            sseConfig: MCPServerSseConfig,
-            requestID: JSONRPCID,
-            requestBody: Data?,
-            lastEventID: String?,
-            retryMilliseconds: Int
-        ) async throws -> JSONRPCMessage {
-            let sessionConfig = URLSessionConfiguration.default
-            sessionConfig.timeoutIntervalForRequest = .infinity
-            sessionConfig.timeoutIntervalForResource = .infinity
-
-            let session = URLSession(configuration: sessionConfig)
-            var request = URLRequest(url: endpointURL)
-            request.httpMethod = requestBody == nil ? "GET" : "POST"
-
-            if let requestBody {
-                request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-                configureSSEPOSTRequest(&request, sseConfig: sseConfig)
-                request.httpBody = requestBody
-            } else {
-                configureSSEGETRequest(&request, sseConfig: sseConfig, lastEventID: lastEventID)
-            }
-
-            let (asyncBytes, response) = try await session.bytes(for: request)
-            guard let httpResponse = response as? HTTPURLResponse else {
-                throw MCPServerProxyError.communicationError("Invalid HTTP response")
-            }
-
-            if let updatedSessionID = httpResponse.value(forHTTPHeaderField: "Mcp-Session-Id") {
-                sessionID = updatedSessionID
-            }
-
-            let contentType = httpResponse.value(forHTTPHeaderField: "Content-Type")?.lowercased() ?? ""
-            switch httpResponse.statusCode {
-            case 200:
-                if contentType.contains("application/json") {
-                    var data = Data()
-                    for try await byte in asyncBytes {
-                        data.append(byte)
-                    }
-
-                    guard let responseMessage = try responseMessage(for: requestID, from: data) else {
-                        throw MCPServerProxyError.communicationError("HTTP 200 did not include JSON-RPC response for request \(requestID.stringValue)")
-                    }
-                    return responseMessage
-                }
-
-                if contentType.contains("text/event-stream") {
-                    return try await readStreamableSSEApple(
-                        asyncBytes: asyncBytes,
-                        endpointURL: endpointURL,
-                        sseConfig: sseConfig,
-                        requestID: requestID,
-                        lastEventID: lastEventID,
-                        retryMilliseconds: retryMilliseconds
-                    )
-                }
-
-                throw MCPServerProxyError.communicationError("Unsupported response content type: \(contentType)")
-            case 202:
-                throw MCPServerProxyError.communicationError("Unexpected HTTP 202 for request \(requestID.stringValue)")
-            default:
-                var data = Data()
-                for try await byte in asyncBytes {
-                    data.append(byte)
-                }
-                let responseBody = String(data: data, encoding: .utf8)?.trimmingCharacters(in: .whitespacesAndNewlines) ?? ""
-                let details = responseBody.isEmpty ? "" : ": \(responseBody)"
-                throw MCPServerProxyError.communicationError("HTTP error \(httpResponse.statusCode)\(details)")
-            }
-        }
-
-        @available(macOS 12.0, iOS 15.0, tvOS 15.0, watchOS 8.0, macCatalyst 15.0, *)
-        // swiftlint:disable:next function_parameter_count
-        private func readStreamableSSEApple(
-            asyncBytes: URLSession.AsyncBytes,
-            endpointURL: URL,
-            sseConfig: MCPServerSseConfig,
-            requestID: JSONRPCID,
-            lastEventID: String?,
-            retryMilliseconds: Int
-        ) async throws -> JSONRPCMessage {
-            var latestEventID = lastEventID
-            var retryMilliseconds = retryMilliseconds
-
-            do {
-                for try await message in asyncBytes.lines.sseMessages() {
-                    if let id = message.id {
-                        latestEventID = id
-                    }
-                    if let retry = message.retry {
-                        retryMilliseconds = retry
-                    }
-
-                    if message.data.isEmpty {
-                        continue
-                    }
-
-                    if let jsonData = message.data.data(using: .utf8),
-                       let decoded = try? JSONDecoder().decode(JSONRPCMessage.self, from: jsonData),
-                       decoded.id == requestID {
-                        switch decoded {
-                        case .response, .errorResponse:
-                            return decoded
-                        case .request, .notification:
-                            break
-                        }
-                    }
-
-                    await processIncomingMessage(event: message.event, data: message.data)
-                }
-            } catch is CancellationError {
-                throw MCPServerProxyError.communicationError("Request stream cancelled before response was received")
-            } catch {
-                if let latestEventID {
-                    try await Task.sleep(nanoseconds: UInt64(retryMilliseconds) * 1_000_000)
-                    return try await streamableRequestResponseApple(
-                        endpointURL: endpointURL,
-                        sseConfig: sseConfig,
-                        requestID: requestID,
-                        requestBody: nil,
-                        lastEventID: latestEventID,
-                        retryMilliseconds: retryMilliseconds
-                    )
-                }
-                throw MCPServerProxyError.communicationError(error.localizedDescription)
-            }
-
-            if let latestEventID {
-                try await Task.sleep(nanoseconds: UInt64(retryMilliseconds) * 1_000_000)
-                return try await streamableRequestResponseApple(
-                    endpointURL: endpointURL,
-                    sseConfig: sseConfig,
-                    requestID: requestID,
-                    requestBody: nil,
-                    lastEventID: latestEventID,
-                    retryMilliseconds: retryMilliseconds
-                )
-            }
-
-            throw MCPServerProxyError.communicationError("SSE stream closed before response was received")
-        }
-    #endif
-
-    #if os(Linux)
-        private func sendStreamableRequestLinux(_ message: JSONRPCMessage, sseConfig: MCPServerSseConfig) async throws -> JSONRPCMessage {
-            guard let requestID = message.id else {
-                throw MCPServerProxyError.communicationError("Message must have an ID")
-            }
-            guard let endpointURL = endpointURL ?? (isStreamableMCPURL(sseConfig.url) ? sseConfig.url : nil) else {
-                throw MCPServerProxyError.communicationError("Not connected to server")
-            }
-
-            let encoder = JSONEncoder()
-            encoder.outputFormatting = [.sortedKeys]
-            let requestBody = try encoder.encode(message)
-
-            return try await streamableRequestResponseLinux(
-                endpointURL: endpointURL,
-                sseConfig: sseConfig,
-                requestID: requestID,
-                requestBody: requestBody,
-                lastEventID: nil,
-                retryMilliseconds: 1000
-            )
-        }
-
-        // swiftlint:disable:next function_parameter_count
-        private func streamableRequestResponseLinux(
-            endpointURL: URL,
-            sseConfig: MCPServerSseConfig,
-            requestID: JSONRPCID,
-            requestBody: Data?,
-            lastEventID: String?,
-            retryMilliseconds: Int
-        ) async throws -> JSONRPCMessage {
-            let delegate = SSEStreamingDelegate { _ in }
-            let session = URLSession(configuration: .ephemeral, delegate: delegate, delegateQueue: nil)
-
-            var request = URLRequest(url: endpointURL)
-            request.httpMethod = requestBody == nil ? "GET" : "POST"
-
-            if let requestBody {
-                request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-                request.httpBody = requestBody
-                configureSSEPOSTRequest(&request, sseConfig: sseConfig)
-            } else {
-                configureSSEGETRequest(&request, sseConfig: sseConfig, lastEventID: lastEventID)
-            }
-
-            let task = session.dataTask(with: request)
-            task.resume()
-
-            let response = await delegate.waitForResponse()
-            guard let httpResponse = response as? HTTPURLResponse else {
-                throw MCPServerProxyError.communicationError("Invalid HTTP response")
-            }
-
-            if let updatedSessionID = httpResponse.value(forHTTPHeaderField: "Mcp-Session-Id") {
-                sessionID = updatedSessionID
-            }
-
-            let contentType = httpResponse.value(forHTTPHeaderField: "Content-Type")?.lowercased() ?? ""
-
-            switch httpResponse.statusCode {
-            case 200:
-                if contentType.contains("application/json") {
-                    var bodyLines: [String] = []
-                    for await line in delegate.lines {
-                        bodyLines.append(line)
-                    }
-
-                    let body = bodyLines.joined(separator: "\n")
-                    guard let responseMessage = try responseMessage(for: requestID, from: Data(body.utf8)) else {
-                        throw MCPServerProxyError.communicationError("HTTP 200 did not include JSON-RPC response for request \(requestID.stringValue)")
-                    }
-                    return responseMessage
-                }
-
-                if contentType.contains("text/event-stream") {
-                    return try await readStreamableSSELinux(
-                        delegate: delegate,
-                        endpointURL: endpointURL,
-                        sseConfig: sseConfig,
-                        requestID: requestID,
-                        lastEventID: lastEventID,
-                        retryMilliseconds: retryMilliseconds
-                    )
-                }
-
-                throw MCPServerProxyError.communicationError("Unsupported response content type: \(contentType)")
-            case 202:
-                throw MCPServerProxyError.communicationError("Unexpected HTTP 202 for request \(requestID.stringValue)")
-            default:
-                var bodyLines: [String] = []
-                for await line in delegate.lines {
-                    bodyLines.append(line)
-                }
-
-                let responseBody = bodyLines.joined(separator: "\n").trimmingCharacters(in: .whitespacesAndNewlines)
-                let details = responseBody.isEmpty ? "" : ": \(responseBody)"
-                throw MCPServerProxyError.communicationError("HTTP error \(httpResponse.statusCode)\(details)")
-            }
-        }
-
-        // swiftlint:disable:next function_parameter_count
-        private func readStreamableSSELinux(
-            delegate: SSEStreamingDelegate,
-            endpointURL: URL,
-            sseConfig: MCPServerSseConfig,
-            requestID: JSONRPCID,
-            lastEventID: String?,
-            retryMilliseconds: Int
-        ) async throws -> JSONRPCMessage {
-            var latestEventID = lastEventID
-            var retryMilliseconds = retryMilliseconds
-
-            for try await message in delegate.lines.sseMessages() {
-                if let id = message.id {
-                    latestEventID = id
-                }
-                if let retry = message.retry {
-                    retryMilliseconds = retry
-                }
-
-                if message.data.isEmpty {
-                    continue
-                }
-
-                if let jsonData = message.data.data(using: .utf8),
-                   let decoded = try? JSONDecoder().decode(JSONRPCMessage.self, from: jsonData),
-                   decoded.id == requestID {
-                    switch decoded {
-                    case .response, .errorResponse:
-                        return decoded
-                    case .request, .notification:
-                        break
-                    }
-                }
-
-                await processIncomingMessage(event: message.event, data: message.data)
-            }
-
-            if let latestEventID {
-                try await Task.sleep(nanoseconds: UInt64(retryMilliseconds) * 1_000_000)
-                return try await streamableRequestResponseLinux(
-                    endpointURL: endpointURL,
-                    sseConfig: sseConfig,
-                    requestID: requestID,
-                    requestBody: nil,
-                    lastEventID: latestEventID,
-                    retryMilliseconds: retryMilliseconds
-                )
-            }
-
-            if let completionError = delegate.completionError {
-                throw MCPServerProxyError.communicationError(completionError.localizedDescription)
-            }
-
-            throw MCPServerProxyError.communicationError("SSE stream closed before response was received")
-        }
-    #endif
-
-    // MARK: - SSE Connection (Apple platforms)
-
-    #if !os(Linux)
-        @available(macOS 12.0, iOS 15.0, tvOS 15.0, watchOS 8.0, macCatalyst 15.0, *)
-        private func connectSSEApple(sseConfig: MCPServerSseConfig, clientName: String, clientVersion: String) async throws {
-            let isStreamableMCP = isStreamableMCPURL(sseConfig.url)
-            if isStreamableMCP {
-                endpointURL = sseConfig.url
-                try await initialize(clientName: clientName, clientVersion: clientVersion)
-                startStreamableGeneralSSEApple(sseConfig: sseConfig)
-                return
-            }
-
-            let sessionConfig = URLSessionConfiguration.default
-            sessionConfig.timeoutIntervalForRequest = .infinity
-            sessionConfig.timeoutIntervalForResource = .infinity
-
-            let session = URLSession(configuration: sessionConfig)
-            var request = URLRequest(url: sseConfig.url)
-            request.httpMethod = "GET"
-            configureSSEGETRequest(&request, sseConfig: sseConfig)
-
-            streamTask = Task {
-                do {
-                    let (asyncBytes, response) = try await session.bytes(for: request)
-                    self.handleSSEResponse(response, sseConfig: sseConfig, isStreamableMCP: isStreamableMCP)
-
-                    for try await message in asyncBytes.lines.sseMessages() {
-                        await self.processIncomingMessage(event: message.event, data: message.data)
-                    }
-                    self.handleStreamTermination(
-                        MCPServerProxyError.communicationError(
-                            "SSE stream closed by server before response was received"
-                        )
-                    )
-                } catch is CancellationError {
-                    // Pending requests are cancelled in disconnect().
-                } catch {
-                    self.logger.error("[MCP DEBUG] SSE stream error: \(error)")
-                    self.handleStreamTermination(error)
-                }
-            }
-
-            try await waitForEndpointIfNeeded(isStreamableMCP: isStreamableMCP)
-            try await initialize(clientName: clientName, clientVersion: clientVersion)
-        }
-    #endif
-
-    // MARK: - SSE Connection (Linux)
-
-    #if os(Linux)
-        private func connectSSELinux(sseConfig: MCPServerSseConfig, clientName: String, clientVersion: String) async throws {
-            let isStreamableMCP = isStreamableMCPURL(sseConfig.url)
-            if isStreamableMCP {
-                endpointURL = sseConfig.url
-                try await initialize(clientName: clientName, clientVersion: clientVersion)
-                startStreamableGeneralSSELinux(sseConfig: sseConfig)
-                return
-            }
-
-            let sessionConfig = URLSessionConfiguration.default
-
-            var request = URLRequest(url: sseConfig.url)
-            request.httpMethod = "GET"
-            configureSSEGETRequest(&request, sseConfig: sseConfig)
-
-            // Use a streaming delegate since URLSession.bytes is unavailable on Linux
-            let proxy = self
-            let delegate = SSEStreamingDelegate { response in
-                Task {
-                    await proxy.handleSSEResponse(response, sseConfig: sseConfig, isStreamableMCP: isStreamableMCP)
-                }
-            }
-
-            let session = URLSession(configuration: sessionConfig, delegate: delegate, delegateQueue: nil)
-            let task = session.dataTask(with: request)
-            task.resume()
-
-            streamTask = Task {
-                do {
-                    for try await message in delegate.lines.sseMessages() {
-                        await self.processIncomingMessage(event: message.event, data: message.data)
-                    }
-                    self.handleStreamTermination(
-                        MCPServerProxyError.communicationError(
-                            "SSE stream closed by server before response was received"
-                        )
-                    )
-                } catch is CancellationError {
-                    task.cancel()
-                } catch {
-                    self.logger.error("[MCP DEBUG] SSE stream error: \(error)")
-                    self.handleStreamTermination(error)
-                }
-            }
-
-            try await waitForEndpointIfNeeded(isStreamableMCP: isStreamableMCP)
-            try await initialize(clientName: clientName, clientVersion: clientVersion)
-        }
-    #endif
-
-    #if !os(Linux)
-        @available(macOS 12.0, iOS 15.0, tvOS 15.0, watchOS 8.0, macCatalyst 15.0, *)
-        private func startStreamableGeneralSSEApple(sseConfig: MCPServerSseConfig) {
-            streamTask = Task {
-                var lastEventID: String?
-                var retryMilliseconds = 1000
-
-                while !self.isDisconnecting {
-                    do {
-                        let sessionConfig = URLSessionConfiguration.default
-                        sessionConfig.timeoutIntervalForRequest = .infinity
-                        sessionConfig.timeoutIntervalForResource = .infinity
-
-                        let session = URLSession(configuration: sessionConfig)
-                        var request = URLRequest(url: sseConfig.url)
-                        request.httpMethod = "GET"
-                        self.configureSSEGETRequest(&request, sseConfig: sseConfig, lastEventID: lastEventID)
-
-                        let (asyncBytes, response) = try await session.bytes(for: request)
-                        self.handleSSEResponse(response, sseConfig: sseConfig, isStreamableMCP: true)
-
-                        for try await message in asyncBytes.lines.sseMessages() {
-                            if let id = message.id {
-                                lastEventID = id
-                            }
-                            if let retry = message.retry {
-                                retryMilliseconds = retry
-                            }
-                            await self.processIncomingMessage(event: message.event, data: message.data)
-                        }
-                    } catch is CancellationError {
-                        return
-                    } catch {
-                        if self.isDisconnecting {
-                            return
-                        }
-                        self.logger.error("[MCP DEBUG] Streamable general SSE stream error: \(error)")
-                    }
-
-                    if self.isDisconnecting {
-                        return
-                    }
-
-                    do {
-                        try await Task.sleep(nanoseconds: UInt64(retryMilliseconds) * 1_000_000)
-                    } catch {
-                        return
-                    }
-                }
-            }
-        }
-    #endif
-
-    #if os(Linux)
-        private func startStreamableGeneralSSELinux(sseConfig: MCPServerSseConfig) {
-            streamTask = Task {
-                var lastEventID: String?
-                var retryMilliseconds = 1000
-
-                while !self.isDisconnecting {
-                    let sessionConfig = URLSessionConfiguration.default
-                    let proxy = self
-                    let delegate = SSEStreamingDelegate { response in
-                        Task {
-                            await proxy.handleSSEResponse(response, sseConfig: sseConfig, isStreamableMCP: true)
-                        }
-                    }
-
-                    var request = URLRequest(url: sseConfig.url)
-                    request.httpMethod = "GET"
-                    self.configureSSEGETRequest(&request, sseConfig: sseConfig, lastEventID: lastEventID)
-
-                    let session = URLSession(configuration: sessionConfig, delegate: delegate, delegateQueue: nil)
-                    let task = session.dataTask(with: request)
-                    task.resume()
-
-                    do {
-                        for try await message in delegate.lines.sseMessages() {
-                            if let id = message.id {
-                                lastEventID = id
-                            }
-                            if let retry = message.retry {
-                                retryMilliseconds = retry
-                            }
-                            await self.processIncomingMessage(event: message.event, data: message.data)
-                        }
-                    } catch is CancellationError {
-                        task.cancel()
-                        return
-                    } catch {
-                        if self.isDisconnecting {
-                            task.cancel()
-                            return
-                        }
-                        self.logger.error("[MCP DEBUG] Streamable general SSE stream error: \(error)")
-                    }
-
-                    if self.isDisconnecting {
-                        task.cancel()
-                        return
-                    }
-
-                    do {
-                        try await Task.sleep(nanoseconds: UInt64(retryMilliseconds) * 1_000_000)
-                    } catch {
-                        task.cancel()
-                        return
-                    }
-                }
-            }
-        }
-    #endif
-
-    // MARK: - Shared SSE Helpers
-
-    private func handleSSEResponse(_ response: URLResponse, sseConfig: MCPServerSseConfig, isStreamableMCP: Bool) {
-        if let httpResponse = response as? HTTPURLResponse {
-            sessionID = httpResponse.value(forHTTPHeaderField: "Mcp-Session-Id")
-            if isStreamableMCP {
-                endpointURL = sseConfig.url
-            } else if let sessionID,
-                      let endpoint = messageEndpointURL(baseURL: sseConfig.url, sessionId: sessionID) {
-                endpointURL = endpoint
-            }
-        }
-    }
-
-    private func waitForEndpointIfNeeded(isStreamableMCP: Bool) async throws {
-        if !isStreamableMCP && endpointURL == nil {
-            _ = try await withCheckedThrowingContinuation { (continuation: CheckedContinuation<URL, Error>) in
-                self.endpointContinuation = continuation
-
-                Task {
-                    try await Task.sleep(nanoseconds: 10_000_000_000)
-                    if let cont = self.endpointContinuation {
-                        self.endpointContinuation = nil
-                        cont.resume(throwing: MCPServerProxyError.communicationError("Timeout waiting for endpoint URL"))
-                    }
-                }
-            }
-        }
-    }
-
-    private func messageEndpointURL(baseURL: URL, sessionId: String) -> URL? {
-        guard var components = URLComponents(url: baseURL, resolvingAgainstBaseURL: false) else {
-            return nil
-        }
-        components.path = "/messages/\(sessionId)"
-        return components.url
-    }
 }
 
-private extension JSONRPCID {
-    var stringValue: String {
+extension JSONRPCID {
+    internal var stringValue: String {
         switch self {
         case .int(let value):
             return String(value)
